Rename typoed transformer module file
