+++ minitorch/__init__.py
@@ -26,6 +26,6 @@ from .tensor_ops import *  # noqa: F401,F403
 from .testing import MathTest, MathTestVariable  # type: ignore # noqa: F401,F403
 
 # ASSIGNMENT 2
-from .modules_transfomer import *
+from .transformer import *
 
 version = "0.4"
